preparing for refactor
