+++ fastmap/utils.py
